Impl process to retrieve latest articles
Created a new process to retrieve the latest artices since the last run
of the app.

This process is very similar to the original process that retrieves all
the existing articles but uses the "since" value from Pocket to only
retrieve articles since a point in time.

The new articles are added to the existing db of articles.